Move spark2 JMH tests to spark directory and update jmh.gradle and build.gradle to allow for running on spark3 and spark2
