+++ spark/src/jmh/java/org/apache/iceberg/spark/source/avro/IcebergSourceNestedAvroDataReadBenchmark.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iceberg.spark.source.parquet;
+package org.apache.iceberg.spark.source.avro;
 
 import java.io.IOException;
 import java.util.Map;
@@ -31,31 +31,28 @@ import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.TearDown;
 import org.openjdk.jmh.annotations.Threads;
 
+import static org.apache.iceberg.TableProperties.DEFAULT_FILE_FORMAT;
 import static org.apache.iceberg.TableProperties.SPLIT_OPEN_FILE_COST;
 import static org.apache.spark.sql.functions.expr;
 import static org.apache.spark.sql.functions.lit;
 import static org.apache.spark.sql.functions.struct;
 
+
 /**
- * A benchmark that evaluates the file skipping capabilities in the Spark data source for Iceberg.
+ * A benchmark that evaluates the performance of reading Avro data with a flat schema
+ * using Iceberg and the built-in file source in Spark.
  *
- * This class uses a dataset with nested data, where the records are clustered according to the
- * column used in the filter predicate.
- *
- * The performance is compared to the built-in file source in Spark.
- *
- * To run this benchmark:
+ * To run this benchmark for either spark-2 or spark-3:
  * <code>
- *   ./gradlew :iceberg-spark2:jmh
- *       -PjmhIncludeRegex=IcebergSourceNestedParquetDataFilterBenchmark
- *       -PjmhOutputPath=benchmark/iceberg-source-nested-parquet-data-filter-benchmark-result.txt
+ *   ./gradlew :iceberg-spark[2|3]:jmh
+ *       -PjmhIncludeRegex=IcebergSourceNestedAvroDataReadBenchmark
+ *       -PjmhOutputPath=benchmark/iceberg-source-nested-avro-data-read-benchmark-result.txt
  * </code>
  */
-public class IcebergSourceNestedParquetDataFilterBenchmark extends IcebergSourceNestedDataBenchmark {
+public class IcebergSourceNestedAvroDataReadBenchmark extends IcebergSourceNestedDataBenchmark {
 
-  private static final String FILTER_COND = "nested.col3 == 0";
-  private static final int NUM_FILES = 500;
-  private static final int NUM_ROWS = 10000;
+  private static final int NUM_FILES = 10;
+  private static final int NUM_ROWS = 1000000;
 
   @Setup
   public void setupBenchmark() {
@@ -71,51 +68,65 @@ public class IcebergSourceNestedParquetDataFilterBenchmark extends IcebergSource
 
   @Benchmark
   @Threads(1)
-  public void readWithFilterIceberg() {
+  public void readIceberg() {
     Map<String, String> tableProperties = Maps.newHashMap();
     tableProperties.put(SPLIT_OPEN_FILE_COST, Integer.toString(128 * 1024 * 1024));
     withTableProperties(tableProperties, () -> {
       String tableLocation = table().location();
-      Dataset<Row> df = spark().read().format("iceberg").load(tableLocation).filter(FILTER_COND);
+      Dataset<Row> df = spark().read().format("iceberg").load(tableLocation);
       materialize(df);
     });
   }
 
   @Benchmark
   @Threads(1)
-  public void readWithFilterFileSourceVectorized() {
+  public void readFileSource() {
     Map<String, String> conf = Maps.newHashMap();
-    conf.put(SQLConf.PARQUET_VECTORIZED_READER_ENABLED().key(), "true");
     conf.put(SQLConf.FILES_OPEN_COST_IN_BYTES().key(), Integer.toString(128 * 1024 * 1024));
     withSQLConf(conf, () -> {
-      Dataset<Row> df = spark().read().parquet(dataLocation()).filter(FILTER_COND);
+      Dataset<Row> df = spark().read().format("avro").load(dataLocation());
       materialize(df);
     });
   }
 
   @Benchmark
   @Threads(1)
-  public void readWithFilterFileSourceNonVectorized() {
+  public void readWithProjectionIceberg() {
+    Map<String, String> tableProperties = Maps.newHashMap();
+    tableProperties.put(SPLIT_OPEN_FILE_COST, Integer.toString(128 * 1024 * 1024));
+    withTableProperties(tableProperties, () -> {
+      String tableLocation = table().location();
+      Dataset<Row> df = spark().read().format("iceberg").load(tableLocation).select("nested.col3");
+      materialize(df);
+    });
+  }
+
+  @Benchmark
+  @Threads(1)
+  public void readWithProjectionFileSource() {
     Map<String, String> conf = Maps.newHashMap();
-    conf.put(SQLConf.PARQUET_VECTORIZED_READER_ENABLED().key(), "false");
     conf.put(SQLConf.FILES_OPEN_COST_IN_BYTES().key(), Integer.toString(128 * 1024 * 1024));
     withSQLConf(conf, () -> {
-      Dataset<Row> df = spark().read().parquet(dataLocation()).filter(FILTER_COND);
+      Dataset<Row> df = spark().read().format("avro").load(dataLocation()).select("nested.col3");
       materialize(df);
     });
   }
 
   private void appendData() {
-    for (int fileNum = 1; fileNum <= NUM_FILES; fileNum++) {
-      Dataset<Row> df = spark().range(NUM_ROWS)
-          .withColumn(
-              "nested",
-              struct(
-                  expr("CAST(id AS string) AS col1"),
-                  expr("CAST(id AS double) AS col2"),
-                  lit(fileNum).cast("long").as("col3")
-              ));
-      appendAsFile(df);
-    }
+    Map<String, String> tableProperties = Maps.newHashMap();
+    tableProperties.put(DEFAULT_FILE_FORMAT, "avro");
+    withTableProperties(tableProperties, () -> {
+      for (int fileNum = 1; fileNum <= NUM_FILES; fileNum++) {
+        Dataset<Row> df = spark().range(NUM_ROWS)
+            .withColumn(
+                "nested",
+                struct(
+                    expr("CAST(id AS string) AS col1"),
+                    expr("CAST(id AS double) AS col2"),
+                    lit(fileNum).cast("long").as("col3")
+                ));
+        appendAsFile(df);
+      }
+    });
   }
 }
