+++ spark/src/jmh/java/org/apache/iceberg/spark/source/parquet/IcebergSourceNestedListParquetDataWriteBenchmark.java
@@ -22,16 +22,18 @@ package org.apache.iceberg.spark.source.parquet;
 import java.io.IOException;
 import java.util.Map;
 import org.apache.iceberg.relocated.com.google.common.collect.Maps;
-import org.apache.iceberg.spark.source.IcebergSourceNestedDataBenchmark;
+import org.apache.iceberg.spark.source.IcebergSourceNestedListDataBenchmark;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
 import org.apache.spark.sql.SaveMode;
 import org.apache.spark.sql.internal.SQLConf;
 import org.openjdk.jmh.annotations.Benchmark;
+import org.openjdk.jmh.annotations.Param;
 import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.TearDown;
 import org.openjdk.jmh.annotations.Threads;
 
+import static org.apache.spark.sql.functions.array_repeat;
 import static org.apache.spark.sql.functions.expr;
 import static org.apache.spark.sql.functions.struct;
 
@@ -39,16 +41,14 @@ import static org.apache.spark.sql.functions.struct;
  * A benchmark that evaluates the performance of writing nested Parquet data using Iceberg
  * and the built-in file source in Spark.
  *
- * To run this benchmark:
+ * To run this benchmark for either spark-2 or spark-3:
  * <code>
- *   ./gradlew :iceberg-spark2:jmh
- *       -PjmhIncludeRegex=IcebergSourceNestedParquetDataWriteBenchmark
- *       -PjmhOutputPath=benchmark/iceberg-source-nested-parquet-data-write-benchmark-result.txt
+ *   ./gradlew :iceberg-spark[2|3]:jmh
+ *       -PjmhIncludeRegex=IcebergSourceNestedListParquetDataWriteBenchmark
+ *       -PjmhOutputPath=benchmark/iceberg-source-nested-list-parquet-data-write-benchmark-result.txt
  * </code>
  */
-public class IcebergSourceNestedParquetDataWriteBenchmark extends IcebergSourceNestedDataBenchmark {
-
-  private static final int NUM_ROWS = 5000000;
+public class IcebergSourceNestedListParquetDataWriteBenchmark extends IcebergSourceNestedListDataBenchmark {
 
   @Setup
   public void setupBenchmark() {
@@ -60,6 +60,9 @@ public class IcebergSourceNestedParquetDataWriteBenchmark extends IcebergSourceN
     tearDownSpark();
     cleanupFiles();
   }
+
+  @Param({"2000", "20000"})
+  private int numRows;
 
   @Benchmark
   @Threads(1)
@@ -77,14 +80,10 @@ public class IcebergSourceNestedParquetDataWriteBenchmark extends IcebergSourceN
   }
 
   private Dataset<Row> benchmarkData() {
-    return spark().range(NUM_ROWS)
-        .withColumn(
-            "nested",
-            struct(
-                expr("CAST(id AS string) AS col1"),
-                expr("CAST(id AS double) AS col2"),
-                expr("id AS col3")
-            ))
+    return spark().range(numRows)
+        .withColumn("outerlist", array_repeat(struct(
+            expr("array_repeat(CAST(id AS string), 1000) AS innerlist")),
+            10))
         .coalesce(1);
   }
 }
